+++ src/lib/components/admin/Users.svelte
@@ -10,6 +10,8 @@
 
 	import UserList from './Users/UserList.svelte';
 	import Groups from './Users/Groups.svelte';
+	import Credit from './Users/Credit.svelte';
+	import CreditLog from './Users/CreditLog.svelte';
 	import Checkin from './Users/Checkin.svelte';
 
 	const i18n = getContext('i18n');
@@ -18,7 +20,7 @@
 	$: {
 		const pathParts = $page.url.pathname.split('/');
 		const tabFromPath = pathParts[pathParts.length - 1];
-		selectedTab = ['overview', 'groups', 'checkin'].includes(tabFromPath)
+		selectedTab = ['overview', 'groups', 'credit', 'creditLog', 'checkin'].includes(tabFromPath)
 			? tabFromPath
 			: 'overview';
 	}
@@ -124,7 +126,7 @@
 				{/if}
 			</a>
 
-			{#each [{ id: 'checkin', label: 'Daily Check-in' }] as tab}
+			{#each [{ id: 'credit', label: 'Credit Statistics' }, { id: 'creditLog', label: 'Credit Log' }, { id: 'checkin', label: 'Daily Check-in' }] as tab}
 				<a
 					id={tab.id}
 					href="/admin/users/{tab.id}"
@@ -144,6 +146,10 @@
 				<UserList />
 			{:else if selectedTab === 'groups'}
 				<Groups />
+			{:else if selectedTab === 'credit'}
+				<Credit />
+			{:else if selectedTab === 'creditLog'}
+				<CreditLog />
 			{:else if selectedTab === 'checkin'}
 				<Checkin />
 			{/if}